% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{add_issue}
\alias{add_issue}
\title{Helper function to add identified issues to the list}
\usage{
add_issue(issues, name, issue_message)
}
\arguments{
\item{issues}{A list containing named vectors of issue messages.}

\item{name}{The category name under which the issue message should be added.}

\item{issue_message}{The message describing the issue.}
}
\value{
The updated list of issues with the new issue message appended.
}
\description{
Helper function to add identified issues to the list
}
