% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessSub2.R
\name{addDateColumn}
\alias{addDateColumn}
\title{Adds a date column to a data frame based on a date string extracted from file names.}
\usage{
addDateColumn(df, date_str)
}
\arguments{
\item{df}{A data frame to which the date column will be added.}

\item{date_str}{A string representing a date, extracted from the names of data files.
The function expects this string to be in one of two formats:
- 'yymmddhh' (year, month, day, hour)
- 'yymmdd' (year, month, day)
The function attempts to detect the format automatically based on the string length
and pattern.}
}
\value{
The input data frame with an added 'Date' column. If the date string matches
        one of the expected formats, this column will contain date values corresponding
        to the string. If the format is unrecognized, the column will be filled with NA,
        and a warning will be issued.
}
\description{
Adds a date column to a data frame based on a date string extracted from file names.
}
