% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{calculate_and_print_feedback}
\alias{calculate_and_print_feedback}
\title{Calculate and Print Feedback for Data Filtering}
\usage{
calculate_and_print_feedback(all_data, filtered_data, varname)
}
\arguments{
\item{all_data}{A data frame containing the original, unfiltered dataset.}

\item{filtered_data}{A data frame containing the dataset after filtering has been applied.}

\item{varname}{A character string representing the variable name used for generating output filenames
for feedback reports (e.g., "ndvi" or "cci").}
}
\value{
The function does not return a value but generates two text files:
        - One summarizing the filtering feedback by 'Code'.
        - Another providing detailed filtering feedback by both 'Date' and 'Code'.
        The user is notified via the console about the generation of these files.
}
\description{
Calculate and Print Feedback for Data Filtering
}
\details{
The function first calculates the total number of observations and the number of filtered
         observations for each 'Code' in the dataset. It then computes the number of observations removed
         and the percentage of observations removed for each 'Code'. This summary is saved to a text file
         named after the variable name with "_feedback.txt" suffix.

         Additionally, the function provides a more detailed feedback report grouped by both 'Date' and 'Code',
         calculating similar statistics. This detailed report is saved to a text file with "_detailed_feedback.txt"
         suffix, allowing for a more granular analysis of the filtering process.

         The function is designed to assist in evaluating the impact of filtering criteria on the dataset,
         offering insights into how data reduction varies across different 'Code' categories and specific dates.
}
\note{
The generated feedback files are saved in the current working directory. Users should ensure
      appropriate permissions are available for file creation and that the working directory is
      set to a desired path before running the function.
}
