% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_data}
\alias{check_data}
\title{Comprehensive Data Check and Issue Logging Function}
\usage{
check_data(df, variable)
}
\arguments{
\item{df}{A list containing data frames to be checked.}

\item{variable}{The name of the variable being checked, used to tailor the checks for specific data types.}
}
\value{
Logical value indicating whether the data passed the checks without issues (TRUE) or if issues were found (FALSE).
}
\description{
Comprehensive Data Check and Issue Logging Function
}
\details{
The function operates in several steps:
         1. It applies two sets of checks: `check_df1` for general data frame issues and `check_df2` for issues
            in data frames excluding index data.
         2. Issues are collected and aggregated across all data frames.
         3. Identified issues are logged into a "data_check_issues.txt" file for further action.
         4. The function also prints warnings in the R console for each data frame with issues, summarizing the problems found.

         This process ensures a comprehensive review of the data, aiding in the identification and correction
         of potential data quality issues before further analysis.
}
\note{
The use of verbose warnings and issue logging provides a dual approach to informing the user
      of potential problems: immediate feedback via the console and a permanent record in a text file.
      This comprehensive feedback mechanism ensures that data issues are not overlooked and can be
      addressed in a systematic manner.
}
