% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_df1}
\alias{check_df1}
\title{Data Frame Structure and Content Validation Function}
\usage{
check_df1(current_df, name)
}
\arguments{
\item{current_df}{The data frame to be checked.}

\item{name}{The name of the data frame, used for identifying it in the issue log.}
}
\value{
A list of identified issues with the data frame structure and content.
}
\description{
Data Frame Structure and Content Validation Function
}
\details{
The function first checks if the data frame exists and is not empty.
         It then compares the data frame's structure against expected structures for different data types.
         This includes checking the number of columns, the names of the columns, and the data types of columns.
         Identified issues are added to a list which is returned by the function.
}
