% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_df2}
\alias{check_df2}
\title{Performs Consistency and Data Integrity Checks on a Data Frame}
\usage{
check_df2(current_df, name, variable)
}
\arguments{
\item{current_df}{The data frame to be validated.}

\item{name}{A string representing the name of the data frame, used for identification in issue logging.}

\item{variable}{A string indicating the type of data being checked, such as 'NDVI' or 'CCI'.}
}
\value{
A list of identified issues, if any, within the data frame.
}
\description{
Performs Consistency and Data Integrity Checks on a Data Frame
}
\details{
The function checks for:
- Sequentiality in the 'index' or 'Sample' column, depending on the data type.
- Correct time and date format in datetime columns.
- Consistency in the 'id' or 'Units' column with the expected variable type.
- Presence of missing values in any of the columns.
- Duplicate rows within the data frame.

It employs a helper function to filter out rows that might contain column names
instead of actual data, ensuring that checks are performed on valid data entries only.
Additionally, it dynamically adjusts checks based on the column structure, catering
specifically to the requirements of NDVI and Chlorophyll data.
}
\note{
This function is part of a larger data validation process and assumes that
the data frame has already undergone preliminary checks for existence and basic structure.
It is designed to be flexible and can be adapted or extended to include additional
checks as needed.
}
