% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{check_naming_pattern}
\alias{check_naming_pattern}
\title{Check naming pattern}
\usage{
check_naming_pattern(file, valid_patterns)
}
\arguments{
\item{file}{The file to check.}

\item{valid_patterns}{A character vector of valid filename patterns.}
}
\value{
TRUE if the file name matches any of the valid patterns, FALSE otherwise.
}
\description{
Check naming pattern
}
