% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_sequential_index}
\alias{check_sequential_index}
\title{Function to Verify Sequentiality in Chlorophyll Data Sample Column}
\usage{
check_sequential_index(df, verbose = FALSE)
}
\arguments{
\item{df}{A dataframe containing chlorophyll data, including a 'Sample' column that needs verification.}

\item{verbose}{A logical parameter that, when set to TRUE, prints a message for each checked segment indicating
successful verification of sequentiality.}
}
\value{
A logical value: TRUE if the 'Sample' numbers are sequential in all segments, and FALSE otherwise.
}
\description{
Function to Verify Sequentiality in Chlorophyll Data Sample Column
}
\details{
The function iterates through the dataframe, identifying segments of data delineated by rows
         with the value "Sample" in the 'Sample' column. Each segment is checked to ensure that the 'Sample'
         numbers are sequential (i.e., increase by exactly 1 between each row). The function can handle
         multiple segments within a single dataframe, making it suitable for datasets that include
         multiple measurement sessions concatenated into one file.

         If any non-sequential 'Sample' values are found within a segment (excluding the "Sample" header rows),
         the function returns FALSE, indicating an issue with the sequentiality of sample numbering.
         Otherwise, it returns TRUE, confirming that all 'Sample' numbers within each segment are sequential.

         This verification is crucial for ensuring data integrity, especially when analyzing time-series
         measurements or when the order of samples is significant for the analysis.
}
\note{
The function uses the 'Sample' column to identify new data segments and to check the sequentiality
      of sample numbers. It is designed to skip over and not count the header rows marked as "Sample".
      Ensure that your data is formatted correctly, with "Sample" used exclusively to denote headers
      for new segments.

      The function suppresses warnings when converting 'Sample' values to numeric, as the presence of
      "Sample" headers will cause conversion warnings. This suppression ensures that the function focuses
      on checking sequentiality without being hindered by expected non-numeric values.
}
