% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Function to get rid of junk from column names}
\usage{
clean_column_names(col_names)
}
\arguments{
\item{col_names}{A vector of column names to clean.}
}
\value{
A vector of cleaned column names.
}
\description{
Function to get rid of junk from column names
}
