% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{create_combined_ndvi}
\alias{create_combined_ndvi}
\title{Create a Combined NDVI DataFrame with Optional Filtering}
\usage{
create_combined_ndvi(
  all_data,
  apply_ndvi_filter = FALSE,
  lower_limit = NULL,
  upper_limit = NULL
)
}
\arguments{
\item{all_data}{A list of data frames containing NDVI data.}

\item{apply_ndvi_filter}{A logical value indicating whether NDVI filtering should be applied. If TRUE, the function filters NDVI values based on the specified lower and upper limits.}

\item{lower_limit}{The lower limit for NDVI filtering. Only NDVI values greater than this limit will be included in the final data frame if filtering is applied.}

\item{upper_limit}{The upper limit for NDVI filtering. Only NDVI values less than this limit will be included in the final data frame if filtering is applied.}
}
\value{
A data frame combining all NDVI data with optional filtering applied.
}
\description{
Create a Combined NDVI DataFrame with Optional Filtering
}
\details{
The function first combines all NDVI data frames from the input list into a single data frame.
         It then selects specific columns to retain in the final combined data frame. The `index` column is removed as it might be redundant.
         The function allows for optional NDVI value filtering. If `apply_ndvi_filter` is set to TRUE and both `lower_limit` and `upper_limit` are provided,
         the function filters the combined data frame to include only the rows where NDVI values fall within the specified range.
         This is useful for focusing on NDVI values of interest by excluding outliers or values outside a specific range.
}
\note{
Assuming `data_list` is a list of NDVI data frames
}
