% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filter_rows_not_in_columns}
\alias{filter_rows_not_in_columns}
\title{Filtering out rows that might inadvertently contain column names as data}
\usage{
filter_rows_not_in_columns(df, required_columns)
}
\arguments{
\item{df}{The data frame to filter.}

\item{required_columns}{A vector of required column names.}
}
\value{
The filtered data frame.
}
\description{
Filtering out rows that might inadvertently contain column names as data
}
