% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{handle_CSV_file}
\alias{handle_CSV_file}
\title{Handle CSV files}
\usage{
handle_CSV_file(file)
}
\arguments{
\item{file}{The CSV file to read.}
}
\value{
A list containing:
        - `data`: A list with the data frame of the CSV file.
        - `files`: A list mapping the file name to the file path.
}
\description{
Handle CSV files
}
