% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{handle_excel_file}
\alias{handle_excel_file}
\title{Handle Excel files}
\usage{
handle_excel_file(file)
}
\arguments{
\item{file}{The Excel file to read.}
}
\value{
A list containing:
        - `data`: A list of data frames for each sheet in the Excel file.
        - `files`: A list mapping sheet names to the file path.
}
\description{
Handle Excel files
}
