% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling_logging.R
\name{handle_general_condition}
\alias{handle_general_condition}
\title{Handler for General Warnings and Errors}
\usage{
handle_general_condition(cond, type = "error", env)
}
\arguments{
\item{cond}{A condition object representing the warning or error to be handled.}

\item{type}{Character string specifying the type of condition: either "error" or "warning". Defaults to "error".}
}
\description{
Handler for General Warnings and Errors
}
\details{
This function handles general warnings and errors by logging the message using `log_general_issue`.
         If the condition is an error, it stops execution by calling `stop` with the message.
         If the condition is a warning, it re-issues the warning using `warning`.
}
