% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{handle_text_file}
\alias{handle_text_file}
\title{Handle text files}
\usage{
handle_text_file(file, skip_rows = 4)
}
\arguments{
\item{file}{The text file to read.}

\item{skip_rows}{The number of rows to skip when reading the text file. Default is 4.}
}
\value{
A list containing:
        - `data`: A list with the data frame of the text file.
        - `files`: A list mapping the file name to the file path.
}
\description{
Handle text files
}
