% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_directory}
\alias{is_directory}
\title{General-purpose helper functions.
Check if a file is a directory}
\usage{
is_directory(file)
}
\arguments{
\item{file}{The file path to check.}
}
\value{
TRUE if the file is a directory, FALSE otherwise.
}
\description{
General-purpose helper functions.
Check if a file is a directory
}
