% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_index_file}
\alias{is_index_file}
\title{Function to determine if a data frame is from an index file}
\usage{
is_index_file(df_name, data_frame_files)
}
\arguments{
\item{df_name}{The name of the data frame.}

\item{data_frame_files}{A list mapping data frame names to file paths.}
}
\value{
TRUE if the data frame is from an index file, FALSE otherwise.
}
\description{
Function to determine if a data frame is from an index file
}
