% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling_logging.R
\name{log_general_issue}
\alias{log_general_issue}
\title{Logging Function for General Errors and Warnings}
\usage{
log_general_issue(msg, log_file = "error_log.txt")
}
\arguments{
\item{msg}{Character string containing the message to be logged.}

\item{log_file}{Character string specifying the file to which the log message should be written. Defaults to "error_log.txt".}
}
\description{
Logging Function for General Errors and Warnings
}
\details{
This function logs general error and warning messages to a specified log file.
         Each log entry is prepended with a timestamp indicating when the log entry was made.
         The log messages are appended to the existing content of the log file.
}
