% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling_logging.R
\name{log_issue}
\alias{log_issue}
\title{Log issues}
\usage{
log_issue(file, issue)
}
\arguments{
\item{file}{The file with an issue.}

\item{issue}{The issue description.}
}
\description{
Log issues
}
