% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{log_issues_to_file}
\alias{log_issues_to_file}
\title{Helper Function to Log Issues into a Text File}
\usage{
log_issues_to_file(issues, filepath)
}
\arguments{
\item{issues}{A list of named issue vectors to be logged.}

\item{filepath}{A string specifying the path to the file where the issues should be logged.}
}
\description{
Helper Function to Log Issues into a Text File
}
\details{
Each issue is logged with its category name followed by the issue messages.
         If an error occurs during file writing, an error message is printed to the console.
}
