% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{log_skipped_file}
\alias{log_skipped_file}
\title{Log skipped files}
\usage{
log_skipped_file(file, reason)
}
\arguments{
\item{file}{The file that was skipped.}

\item{reason}{The reason why the file was skipped.}
}
\description{
Log skipped files
}
