% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling_logging.R
\name{log_validation_issue}
\alias{log_validation_issue}
\title{Logging Function for Validation Issues}
\usage{
log_validation_issue(msg, log_file = "val_issues.txt")
}
\arguments{
\item{msg}{Character string containing the message to be logged.}

\item{log_file}{Character string specifying the file to which the log message should be written. Defaults to "val_issues.txt".}
}
\description{
Logging Function for Validation Issues
}
\details{
This function logs validation-related messages to a specified log file.
         Each log entry is prepended with a timestamp indicating when the log entry was made.
         The log messages are appended to the existing content of the log file.
}
