% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessMain.R
\name{map_CCI}
\alias{map_CCI}
\title{Map and Transform Elements from Data Files}
\usage{
map_CCI(data)
}
\arguments{
\item{data}{A list containing the raw data frames to be processed.}
}
\value{
A list of transformed data frames, each corresponding to a merged and transformed pair of raw data and index data.
}
\description{
Map and Transform Elements from Data Files
}
\details{
The function follows these steps:
         1. Preprocesses all data frames to standardize their structure and add necessary identifiers.
         2. Merges each data frame with its corresponding index frame based on a common identifier.
         3. Applies a series of transformations to the merged data frames to select specific columns,
            add date columns, rename columns, and convert data types as necessary.

         The preprocessing step involves splitting columns where necessary and assigning unique identifiers
         to ensure that data from different segments can be accurately merged. The merging step combines
         data frames with their indexes to enrich the data with additional context. Finally, the transformation
         step standardizes the format of the data frames, making them ready for analysis.
}
