% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessMain.R
\name{map_NDVI}
\alias{map_NDVI}
\title{Function to select the desired rows (defined in the excel file) from the raw data (txt file)!}
\usage{
map_NDVI(data, data_name, last_df)
}
\arguments{
\item{data}{The raw data frame to be processed.}

\item{data_name}{The name of the data file being processed.}

\item{last_df}{The last data frame for reference.}
}
\value{
A transformed data frame.
}
\description{
Function to select the desired rows (defined in the excel file) from the raw data (txt file)!
}
\details{
This function filters and transforms the data based on specified criteria and merges it with the last data frame.
}
