% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessSub2.R
\name{merge_data_and_index_frames}
\alias{merge_data_and_index_frames}
\title{Merge Data and Index Frames
This function merges data frames from csv_elements with their corresponding index frames from index_elements.
The matching is based on a common 'UniqueID' column present in both sets of data frames. It is designed to work
with data frames named such that the base name of a CSV file matches the name of its corresponding index frame.}
\usage{
merge_data_and_index_frames(csv_elements, index_elements)
}
\arguments{
\item{csv_elements}{A named list of data frames originating from CSV files.}

\item{index_elements}{A named list of index data frames.}
}
\value{
A list of merged data frames, named by the base name of the original CSV files, without the .CSV extension.
}
\description{
Merge Data and Index Frames
This function merges data frames from csv_elements with their corresponding index frames from index_elements.
The matching is based on a common 'UniqueID' column present in both sets of data frames. It is designed to work
with data frames named such that the base name of a CSV file matches the name of its corresponding index frame.
}
