% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_vignette.R
\name{open_vignette}
\alias{open_vignette}
\title{Open the HTML Vignette for VegetationAnalysisToolkit}
\usage{
open_vignette()
}
\value{
Opens the HTML vignette in the default browser.
}
\description{
This function opens the HTML vignette for the VegetationAnalysisToolkit package in your default web browser.
The vignette provides an introduction and examples to help you get started with the package.
}
\examples{
open_vignette()
}
