% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_NDVI}
\alias{plot_NDVI}
\title{Plot NDVI Data}
\usage{
plot_NDVI(df, code, tstart, tend, plot_type)
}
\arguments{
\item{df}{A dataframe with columns 'Date', 'Code', and 'NDVI'.}

\item{code}{The specific code to filter by (optional).}

\item{tstart}{The start date for filtering (optional).}

\item{tend}{The end date for filtering (optional).}

\item{plot_type}{The type of plot to generate: "time_series", "boxplot", or "violin".}
}
\value{
A plotly object of the specified NDVI plot.
}
\description{
This function generates various types of plots for NDVI data including time series, box plots, and violin plots.
It allows for filtering by location codes and date range.
}
\details{
Based on the input data (\code{df}) and variable (\code{code}) and plot settings (\code{plot_type}), a plot is created.
The time range can be adjusted by \code{tstart} and \code{tend}.
}
\examples{
\dontrun{
df <- readxl::read_excel("path/to/your/data/ndvi_2023_filtered.xlsx")

# plot timeseries for all locations and date range
plot_NDVI(df)

# plot boxplot for specified dates and location
plot_NDVI(df, code = "CSTA", tstart = "2023-05-12", tend = "2023-08-18", plot_type = "boxplot")
}
}
