% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessSub1.R
\name{preprocess_all_data_frames}
\alias{preprocess_all_data_frames}
\title{Preprocess All Data Frames with Segmentation
This function preprocesses both data and index data frames by segmenting them
and adding unique identifiers. It is designed to handle data frames where
multiple segments exist within a single data frame, often demarcated by specific
header rows. The function differentiates between data and index files based on
their naming conventions.}
\usage{
preprocess_all_data_frames(data)
}
\arguments{
\item{data}{A list containing data frames to be preprocessed.}
}
\value{
A list with two elements: `data_files_processed` and `index_files_processed`,
        each containing preprocessed data frames with added unique identifiers.
}
\description{
Preprocess All Data Frames with Segmentation
This function preprocesses both data and index data frames by segmenting them
and adding unique identifiers. It is designed to handle data frames where
multiple segments exist within a single data frame, often demarcated by specific
header rows. The function differentiates between data and index files based on
their naming conventions.
}
