% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessMain.R
\name{process_Chlorophyll}
\alias{process_Chlorophyll}
\title{Ensure that each file starts with a header comment that explains the purpose of the file and the functions it contains.
Process Chlorophyll Data for Analysis}
\usage{
process_Chlorophyll(
  folder_path,
  output,
  filename,
  lower_limit,
  upper_limit,
  variable,
  validate
)
}
\arguments{
\item{folder_path}{The directory path where the chlorophyll data files are located.}

\item{output}{The directory path where the processed files will be saved.}

\item{filename}{The base name for the output files.}

\item{lower_limit}{The lower limit for chlorophyll data filtering.}

\item{upper_limit}{The upper limit for chlorophyll data filtering.}

\item{variable}{The target variable name for consistency checks, typically "CCI" for chlorophyll content index.}

\item{validate}{A logical flag to indicate whether data integrity checks should be performed.}
}
\value{
None explicitly, but the processed data is saved as Excel files in the specified output directory.
}
\description{
This function streamlines the process of reading, validating, transforming, and saving chlorophyll data.
             It ensures that the data is in the correct format and structure for further analysis.
}
\details{
The function performs several key operations:
  1. Reads in chlorophyll data from specified folder paths.
  2. Optionally validates the data for integrity and structure.
  3. Applies a series of transformations to standardize and enrich the data.
  4. Saves the transformed data into designated output files.
}
\section{README}{

Guidelines for Preparing Your Input Files:

Ensure your input data files are properly formatted and named according to the following guidelines to
facilitate successful processing:

- **File Naming Convention**: Name your data files to include the date of measurement. Use the format %y%m%d for dates (e.g., 230528 for May 28, 2023) or %y%m%d%h for specific hours (e.g., 23052810 for 10 AM on May 28, 2023). Simple date formats are preferred, such as 230528 or 23052810.

- **Required Data Structure**: Your data has to be in the first sheet and first column of a CSV file. Although the sheet name can be arbitrary, ensure that column names start from the first row.
  Your data should include the following sequence of column name-data pairs in the first column:
  <Sample, Time/Date, Units, Reading, Lat, Lon, DOP, # Sat, Point>. This structure is crucial for the script to correctly interpret and process your data.

- **Index File Requirement**: Accompany your measurement data with an Excel file containing indexes and associated codes for each measurement. Name the tabs within this Excel file after the measurement dates to correlate them directly with the data files.

**Coding and Marking of Data Sites**:
In the provided example (refer to the index file's "position" column), data sites are designated with specific codes:
RA, RF, EA, EF, SZA, SZF, CSUA, CSUF, CSUASK, CSUFSK, CSUAHP, CSUAHPSK, CSRA, CSRF, CSRASK, CSRFSK, CSRAHP, CSRAHPSK, CSTA, CSTF, CSTASK, CSTFSK.

**Supported Site Coding**:
- **POSITION**: F (Front), A (Adjacent), HP (High Point)
- **LANDUSE**: CSU (Crop Standard Use), CSR (Crop Sustainable Rotation), CST (Crop Sustainable Tillage), R (Rural), E (Ecological), SZ (Special Zone)
- **INTERROW**: S (Standard Row), SK (Skipped Row)

It is recommended to manually review your input data files for adherence to these guidelines before proceeding with the script. This preliminary check helps ensure smooth data processing and accurate results.
}

\examples{
# Process chlorophyll data with validation and save the results
\dontrun{
process_Chlorophyll("path/to/data", "path/to/output", "chlorophyll_processed",
0.2, 2.5, "CCI", TRUE)
}

}
