% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_dataframe}
\alias{process_dataframe}
\title{Process and Clean DataFrame}
\usage{
process_dataframe(df)
}
\arguments{
\item{df}{A data frame to be processed.}
}
\value{
A list containing:
        - `current_df`: The processed data frame.
        - `current_df_old`: The original data frame before processing.
}
\description{
Process and Clean DataFrame
}
