% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mc100_chlorophyll.R
\name{process_MC100_Chlorophyll}
\alias{process_MC100_Chlorophyll}
\title{Process Data for MC-100 Chlorophyll Concentration Meter}
\usage{
process_MC100_Chlorophyll(
  folder_path,
  output,
  device_type = "MC100",
  lower_limit,
  upper_limit,
  variable,
  validate,
  split_code = FALSE
)
}
\arguments{
\item{folder_path}{The directory path where the data files are located.}

\item{output}{The directory path where the processed files will be saved.}

\item{device_type}{A character string specifying the type of data ("PPNP" for PlantPen or "MC100" for MC-100).}

\item{lower_limit}{The lower limit for data filtering.}

\item{upper_limit}{The upper limit for data filtering.}

\item{variable}{The target variable name for consistency checks, typically "CCI" for chlorophyll content index.}

\item{validate}{A logical flag to indicate whether data integrity checks should be performed.}

\item{split_code}{Logical. If \code{TRUE}, measurement codes will be split based on a configuration file.}
}
\value{
None explicitly, but the processed data is saved as Excel files in the specified output directory.
}
\description{
This function processes raw data from the MC-100 Chlorophyll Concentration Meter.
}
\details{
Key operations of the function include:
\itemize{
\item Optional, validating the data.
\item Applying a series of transformations and filters to clean the data.
\item Saving the processed data in \code{xlsx} format for further analysis.
\item Optional, splitting the measurement code columns based on your given criteria.
}

For input file formatting details, please refer to the package's vignette.
}
\examples{
\dontrun{
process_MC100_Chlorophyll("path/to/data", "path/to/output", 0.2, 2.5, "CCI", TRUE)
}
}
