% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessMain.R
\name{process_NDVI}
\alias{process_NDVI}
\title{Process NDVI Data for Analysis}
\usage{
process_NDVI(
  folder_path,
  output,
  filename,
  lower_limit,
  upper_limit,
  variable,
  validate
)
}
\arguments{
\item{folder_path}{The directory path where the NDVI data files are located.}

\item{output}{The directory path where the processed files will be saved.}

\item{filename}{The base name for the output files.}

\item{lower_limit}{The lower limit for NDVI data filtering.}

\item{upper_limit}{The upper limit for NDVI data filtering.}

\item{variable}{The target variable name for consistency checks, typically "NDVI".}

\item{validate}{A logical flag to indicate whether data integrity checks should be performed.}
}
\value{
None explicitly, but processed NDVI data is saved as Excel files in the specified output directory.
}
\description{
This function reads, validates, transforms, and saves NDVI data from raw measurement files.
             It filters out extreme values and saves the data in a structured format for further analysis.
}
\details{
The function performs several key operations:
  1. Reads in NDVI data from specified folder paths.
  2. Removes unnecessary columns that are not relevant for analysis.
  3. Optionally validates the data for integrity and structure.
  4. Applies transformations to standardize and enrich the data.
  5. Saves the transformed data into designated output files.
}
\section{README}{

Guidelines for Preparing Your Input Files:

Ensure your input data files are properly formatted and named according to the following guidelines to
facilitate successful processing:

- **File Naming Convention**: Your input file names can be any strings, but for Excel files, sheet names must
  reflect the measurement date in the format %y%m%d (e.g., plantpen_NDVI_230928) to facilitate date calculation.

- **Required Data Structure**: In the Excel file, column names (index, description) must start from the first row.
  The raw data Text file should include columns: <index, time, id, 760 [nm], 635 [nm], NDVI>. The script is
  optimized to skip the first four non-relevant rows.

- **Index File Requirement**: Accompany your NDVI measurements with an Excel file specifying each measurement's location.

**Coding and Marking of Data Sites**:
Data sites are designated with specific codes as seen in the "description" column of the Excel index file.
Supported site coding is detailed in the sample output and includes codes for POSITION, LANDUSE, and INTERROW.

It is recommended to manually review your input data files for adherence to these guidelines before proceeding with
the script. This preliminary check helps ensure smooth data processing and accurate results.
}

\examples{
\dontrun{
process_NDVI("path/to/ndvi_data", "path/to/output", "ndvi_processed", 0.1, 0.9, "NDVI", TRUE)
}

}
