% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_plantpen_ndvi_pri.R
\name{process_PlantPen_NDVI_PRI}
\alias{process_PlantPen_NDVI_PRI}
\title{Process Data for PlantPen NDVI & PRI}
\usage{
process_PlantPen_NDVI_PRI(
  folder_path,
  output,
  device_type = "PPNP",
  lower_limit,
  upper_limit,
  variable,
  validate,
  split_code = FALSE
)
}
\arguments{
\item{folder_path}{The path to the folder containing the input data files.}

\item{output}{The directory where the processed data will be saved.}

\item{device_type}{A character string specifying the type of data ("PPNP" for PlantPen or "MC100" for MC-100).}

\item{lower_limit}{The lower limit for values.}

\item{upper_limit}{The upper limit for values.}

\item{variable}{The variable name to process (e.g. NDVI).}

\item{validate}{A boolean flag indicating whether to validate the input data.}

\item{split_code}{Logical. If \code{TRUE}, the measurement codes will be split based on a configuration file.}
}
\value{
None explicitly, but processed data is saved as Excel files in the specified output directory.
}
\description{
This function reads and processes raw measurement files of the PlantPen NDVI & PRI device.
}
\details{
It performs several key actions:
\itemize{
\item Optional, validating the data.
\item Filtering out values outside of specified limits.
\item Saving the processed data in \code{xlsx} format for further analysis.
\item Optional, splitting the measurement code columns based on your given criteria.
}
}
\section{Input File Guidelines}{

For detailed information on how to format your input files, please refer to the \verb{Project Data Management Guidelines} section
in the package's \code{VegetationAnalysisToolkit} vignette file. These resources provide complete instructions on:

\itemize{
\item \strong{File Naming Convention}: Proper naming of input files to facilitate processing.
\item \strong{Required Data Structure}: The expected structure of raw data files and corresponding index files.
\item \strong{Index File Requirement}: Guidelines for using Excel index files to specify measurement locations.
\item \strong{Split_Code Functionality}: How to create the \code{config} file for your project.
}

It is highly recommended to consult the vignette to ensure your input data files adhere to these
guidelines, which will help avoid processing errors and ensure accurate results.
}

\examples{
\dontrun{
vignette("VegetationAnalysisToolkit")
}
\dontrun{
process_PlantPen_NDVI_PRI("path/to/ndvi_data", "path/to/output", 0.1, 0.9, "NDVI", TRUE)
}
}
