% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReading.R
\name{read_Chlorophyll}
\alias{read_Chlorophyll}
\title{Read Chlorophyll Data with Index files}
\usage{
read_Chlorophyll(folder_path)
}
\arguments{
\item{folder_path}{A character string specifying the path to the folder containing the chlorophyll
data files.}
}
\value{
A list containing data frames of chlorophyll data, organized by file name for CSV files or
sheet name for Excel files.
}
\description{
Read Chlorophyll Data with Index files
}
\details{
The function begins by listing all files in the specified directory. For each file, it checks
if the file is a directory or a temporary file, skipping these. For Excel files, it reads each sheet
separately, ensuring that even if a file is temporarily generated (not starting with "~$"), it's considered
for data extraction. For CSV files, it employs a specific separator (";") and ensures UTF-8 encoding for
compatibility with diverse data formats. The function includes error handling via tryCatch to manage issues
in file reading, logging errors without halting execution.
}
\note{
This function currently assumes specific file extensions (.xlsx for Excel and .CSV for CSV files) and
encoding for CSV files. Future improvements may include dynamic separator handling and adaptive row skipping
to accommodate files with varying structures and formats.

Future Improvements:
- Extend support for additional file formats and separators, moving beyond the current limitations of ";"
  for CSV files and the strict file extension checks.
- Implement adaptive row skipping for CSV files to automatically adjust to files with different header
  structures, enhancing the function's flexibility and user-friendliness.
}
