% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{read_NDVI}
\alias{read_NDVI}
\title{Functions for handling file reading and processing.
Read NDVI Data with Index files}
\usage{
read_NDVI(
  folder_pathway,
  skip_rows = 4,
  valid_patterns = c("plantpen_NDVI_\\\\d{6}\\\\.txt$", "index\\\\.xlsx$")
)
}
\arguments{
\item{folder_pathway}{The path to the folder containing NDVI data and its index files.}

\item{skip_rows}{The number of rows to skip when reading the data files. Default is 4.}

\item{valid_patterns}{A character vector of valid filename patterns to match. Default is c("plantpen_NDVI_\\d{6}\\.txt$", "index\\.xlsx$").}
}
\value{
A list containing:
        - `all_data`: A list of data frames containing NDVI data and index files, organized by file name or sheet name.
        - `data_frame_files`: A list mapping data frame names to their respective file paths.
}
\description{
Functions for handling file reading and processing.
Read NDVI Data with Index files
}
\details{
This function was optimized to read NDVI raw text files generated by instrument X (version Y)
and user-defined Excel files that record measurements. For text files, it assumes a specific format
(tab-separated values) but includes a provision for handling files with varying structures through potential
future enhancements.
}
