% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessSub2.R
\name{renameColumnsByPosition}
\alias{renameColumnsByPosition}
\title{Renames columns of a data frame based on the provided list of new names.}
\usage{
renameColumnsByPosition(df, new_names)
}
\arguments{
\item{df}{A data frame whose columns are to be renamed.}

\item{new_names}{A character vector of new names for the columns of the data frame.
The length of this vector must match the number of columns in the data frame.}
}
\value{
The data frame with its columns renamed if the lengths match,
        otherwise returns NULL as an indication of a mismatch.
}
\description{
Renames columns of a data frame based on the provided list of new names.
}
