% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling_logging.R
\name{safely_read_NDVI}
\alias{safely_read_NDVI}
\title{Functions related to logging and handling errors and warnings, as well as giving feedback to the user.}
\usage{
safely_read_NDVI(folder_path, env)
}
\description{
Functions related to logging and handling errors and warnings, as well as giving feedback to the user.
}
