% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{save_cci}
\alias{save_cci}
\title{Save CCI Data to Excel Files}
\usage{
save_cci(df, output, filename, lower_limit, upper_limit)
}
\arguments{
\item{df}{A list or data frame of CCI data to be processed.}

\item{output}{The directory path where the Excel files will be saved.}

\item{filename}{The base name for the output Excel files, to which "_filtered"
will be appended for the file containing filtered data.}

\item{lower_limit}{The lower limit of acceptable CCI values, below which data
will be excluded from the filtered dataset.}

\item{upper_limit}{The upper limit of acceptable CCI values, above which data
will be excluded from the filtered dataset.}
}
\value{
Similar to `save_ndvi`, the function does not return a value explicitly, it saves two
        Excel files to the specified directory: one with the raw combined CCI data
        and another with the filtered CCI data. It also outputs to the console the
        paths to the saved files and details about the filtering process.
}
\description{
Save CCI Data to Excel Files
}
\details{
Filters CCI data, then saves the raw and filtered data to separate Excel files.
}
