% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{save_ndvi}
\alias{save_ndvi}
\title{Save NDVI Data to Excel Files}
\usage{
save_ndvi(df, output, filename, lower_limit, upper_limit)
}
\arguments{
\item{df}{A list or data frame of NDVI data to be processed.}

\item{output}{The directory path where the Excel files will be saved.}

\item{filename}{The base name for the output Excel files. The function will
append "_filtered" to the filename for the filtered data file.}

\item{lower_limit}{The lower limit of acceptable NDVI values. Rows with NDVI
values below this limit will be filtered out from the final data.}

\item{upper_limit}{The upper limit of acceptable NDVI values. Rows with NDVI
values above this limit will be filtered out from the final data.}
}
\value{
The function does not explicitly return a value but saves two Excel
        files to the specified output directory: one containing the raw
        combined NDVI data and another containing the filtered NDVI data.
        It also prints to the console the location of the saved files and
        information about the filtered data.
}
\description{
Save NDVI Data to Excel Files
}
\details{
The function first combines all the NDVI data frames into a single
         data frame. It then creates a second data frame where it applies
         filters based on the specified NDVI value limits. This allows for
         the exclusion of data points that fall outside the desired NDVI range.
         Both the combined (raw) and filtered data frames are then saved to
         separate Excel files in the specified output directory. The function
         also calculates and prints feedback on the amount of data filtered
         and saves the filtered data frame as a global variable for further use.
}
\note{
The function uses the <<- operator to assign the filtered data frame to
      a global variable named `filtered`. This is intended for convenience
      in further analyses but should be used with caution, especially in
      environments where global variable usage might lead to unintended side
      effects.
}
