% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project_directory.R
\name{setup_project_directory}
\alias{setup_project_directory}
\title{Set Up Project Directory for VegetationAnalysisToolkit}
\usage{
setup_project_directory(project_dir = "VegetationAnalysisToolkit_Project")
}
\arguments{
\item{project_dir}{A character string specifying the name of the project directory.
Defaults to "VegetationAnalysisToolkit_Project".}
}
\value{
The normalized path to the created project directory.
}
\description{
This function sets up an example directory structure for a Vegetation Analysis Toolkit project,
including sample data, example scripts, and documentation files.
}
\details{
This function creates the following directory structure within the project directory:
\itemize{
\item \verb{data/}: Contains sample data files for NDVI and Chlorophyll data.
\item \verb{examples/}: Contains example R scripts.
\item \verb{outputs/}: A folder for storing analysis results.
\item \verb{docs/}: Documentation files like vignettes and the README file.
}

Additionally, it sets up a \code{main.R} script for users to start their analysis workflow,
and generates a \code{README.txt} with instructions for getting started.
}
\examples{
\dontrun{
setup_project_directory("My_Vegetation_Project")
}
}
