% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessSub2.R
\name{splitCodeColumn}
\alias{splitCodeColumn}
\title{Splits the "Code" column into three new columns based on specific patterns.}
\usage{
splitCodeColumn(df)
}
\arguments{
\item{df}{A data frame with a column named "Code" that contains strings to be split.}
}
\value{
A data frame with three new columns ("Position", "Landuse", "InterRow") extracted from the "Code" column.
        If a pattern does not match exactly, the resulting column value will be NA.
}
\description{
Splits the "Code" column into three new columns based on specific patterns.
}
\details{
This function is designed to accurately parse and categorize components of a code string
         into distinct categories based on predefined patterns. The function prioritizes exact matches
         to ensure that only fully matching codes are extracted. In cases where a code component is
         expected but not present in the exact form (e.g., "GERGELY1" should not match any pattern),
         the function will not extract a partial match, aiming to maintain the integrity of the categorized data.
}
