% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessSub1.R
\name{splitColumns}
\alias{splitColumns}
\title{Define a function to split a column with multiple data into separate columns (as many as data we have separated by commas)!}
\usage{
splitColumns(input_data)
}
\arguments{
\item{input_data}{A data frame where the first column contains multiple data points separated by commas.}
}
\value{
A data frame with the first column split into multiple columns based on comma separation,
        with each new column named according to the split parts.
}
\description{
Define a function to split a column with multiple data into separate columns (as many as data we have separated by commas)!
}
