% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegetationAnalysisToolkit-package.R
\docType{package}
\name{VegetationAnalysisToolkit-package}
\alias{VegetationAnalysisToolkit}
\alias{VegetationAnalysisToolkit-package}
\title{VegetationAnalysisToolkit: Functions for Vegetation Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Supports detailed handling of vegetation indexes (such as Chlorophy and NDVI data), offering functions for efficient processing, analysis, and visualization capabilities for comprehensive insights into vegetation health and dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anonymus06/VegetationAnalysisToolkit}
  \item Report bugs at \url{https://github.com/anonymus06/VegetationAnalysisToolkit/issues}
}

}
\author{
\strong{Maintainer}: Levente Czelnai \email{levente.czelnai@gmail.com}

Other contributors:
\itemize{
  \item ChatGPT OpenAI (ChatGPT was used for assistance in developing this package.) [contributor]
}

}
\keyword{internal}
