% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Diagnostic tool for data validation and checks.
Comprehensive Data Check and Issue Logging Function}
\usage{
check_data(df, data_type, data_frame_files)
}
\arguments{
\item{df}{A list containing data frames to be checked.}

\item{variable}{The name of the variable being checked, used to tailor the checks for specific data types.}
}
\value{
A list containing:
        - `is_valid`: Logical value indicating whether the data passed the checks without issues (TRUE) or if issues were found (FALSE).
        - `validation_warnings`: A list of warnings encountered during validation.
}
\description{
Diagnostic tool for data validation and checks.
Comprehensive Data Check and Issue Logging Function
}
\details{
The function operates in several steps:
         1. It applies two sets of checks: `check_df1` for general data frame issues and `check_df2` for issues
            in data frames excluding index data.
         2. Issues are collected and aggregated across all data frames.
         3. The function returns a list indicating whether the data passed the checks without issues and the collected validation warnings.

         This process ensures a comprehensive review of the data, aiding in the identification and correction
         of potential data quality issues before further analysis.
}
