% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_sequential_index}
\alias{check_sequential_index}
\title{Helper function to check if a sequence is sequential}
\usage{
check_sequential_index(df, verbose = FALSE)
}
\arguments{
\item{df}{A data frame containing the sequence.}

\item{verbose}{Logical, if TRUE prints detailed messages.}
}
\value{
TRUE if the sequence is sequential, FALSE otherwise.
}
\description{
Helper function to check if a sequence is sequential
}
