% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{create_combined_ndvi}
\alias{create_combined_ndvi}
\title{Create a Combined NDVI DataFrame with Optional Filtering}
\usage{
create_combined_ndvi(
  all_data,
  apply_ndvi_filter = FALSE,
  lower_limit = NULL,
  upper_limit = NULL
)
}
\arguments{
\item{all_data}{A list of data frames containing NDVI data.}

\item{apply_ndvi_filter}{A logical value indicating whether NDVI filtering should be applied. If TRUE, the function filters NDVI values based on the specified lower and upper limits.}

\item{lower_limit}{The lower limit for NDVI filtering.}

\item{upper_limit}{The upper limit for NDVI filtering.}
}
\value{
A data frame combining all NDVI data with optional filtering applied.
}
\description{
Create a Combined NDVI DataFrame with Optional Filtering
}
