% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_NDVI}
\alias{process_NDVI}
\title{Functions specifically involved in processing data.
Process NDVI Data for Analysis}
\usage{
process_NDVI(
  folder_path,
  output,
  filename,
  lower_limit,
  upper_limit,
  variable,
  validate
)
}
\arguments{
\item{folder_path}{The path to the folder containing the input NDVI data files.}

\item{output}{The directory where the processed data will be saved.}

\item{filename}{The base name for the output files.}

\item{lower_limit}{The lower limit for NDVI values.}

\item{upper_limit}{The upper limit for NDVI values.}

\item{variable}{The variable name to process within the NDVI data.}

\item{validate}{A boolean flag indicating whether to validate the input data.}
}
\value{
None explicitly, but processed NDVI data is saved as Excel files in the specified output directory.
}
\description{
This function reads raw NDVI measumenet files made by XXX, optionally validates the datasets,
filters out extreme values and saves the data in a structured format for further analysis.
}
\section{Input File Guidelines}{

Ensure your input data files are properly formatted and named according to the following guidelines to
facilitate successful processing:

- **File Naming Convention**: Your input file names can be any strings, but for Excel files, sheet names must
  reflect the measurement date in the format %y%m%d (e.g., plantpen_NDVI_230928) to facilitate date calculation.

- **Required Data Structure**: In the Excel file, column names (index, description) must start from the first row.
  The raw data Text file should include columns: <index, time, id, 760 [nm], 635 [nm], NDVI>. The script is
  optimized to skip the first four non-relevant rows.

- **Index File Requirement**: Accompany your NDVI measurements with an Excel file specifying each measurement's location.

**Coding and Marking of Data Sites**:
Data sites are designated with specific codes as seen in the "description" column of the Excel index file.
Supported site coding is detailed in the sample output and includes codes for POSITION, LANDUSE, and INTERROW.

It is recommended to manually review your input data files for adherence to these guidelines before proceeding with
the script. This preliminary check helps ensure smooth data processing and accurate results.
}

\examples{
\dontrun{
process_NDVI("path/to/ndvi_data", "path/to/output", "ndvi_processed", 0.1, 0.9, "NDVI", TRUE)
}

}
