% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{save_cci}
\alias{save_cci}
\title{Save CCI Data to Excel Files}
\usage{
save_cci(df, output, filename, lower_limit, upper_limit)
}
\arguments{
\item{df}{A list or data frame of CCI data to be processed.}

\item{output}{The directory path where the Excel files will be saved.}

\item{filename}{The base name for the output Excel files, to which "_filtered"
will be appended for the file containing filtered data.}

\item{lower_limit}{The lower limit of acceptable CCI values, below which data
will be excluded from the filtered dataset.}

\item{upper_limit}{The upper limit of acceptable CCI values, above which data
will be excluded from the filtered dataset.}
}
\value{
Although the function does not return a value explicitly, it saves two
        Excel files to the specified directory: one with the raw combined CCI data
        and another with the filtered CCI data. It also outputs to the console the
        paths to the saved files and details about the filtering process.
}
\description{
Save CCI Data to Excel Files
}
\details{
The function first combines all the CCI data frames into a single data
         frame. It then applies filtering based on the specified lower and upper
         CCI value limits, creating a second data frame with the filtered data.
         Both the raw (combined) and filtered data frames are saved to separate
         Excel files in the specified output directory. The function also calculates
         and provides feedback on the volume of data filtered. Additionally, it
         saves the filtered data frame as a global variable for easy access in
         subsequent analyses.
}
\note{
Similar to `save_ndvi`, this function uses the <<- operator to assign the
      filtered data frame to a global variable named `filtered`. This practice
      facilitates further analysis but should be approached with caution to avoid
      potential side effects in environments where global variables might lead to
      conflicts or unexpected behavior.
}
